requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai
sentence-transformers
python-dotenv
chromadb
pypdf
fpdf2
sqlalchemy